+++ src/plugin-marklogic/main/uk/co/reecedunn/intellij/plugin/marklogic/rewriter/completion/RewriterCompletionContributor.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2019-2020 Reece H. Dunn
+ * Copyright (C) 2020 Reece H. Dunn
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -13,14 +13,17 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package uk.co.reecedunn.intellij.plugin.xpath.completion.lookup
+package uk.co.reecedunn.intellij.plugin.marklogic.rewriter.completion
 
-class XPathKeywordLookup(
-    lookupString: String,
-    override val insertText: XPathInsertText? = null
-) : XPathLookupElement(lookupString) {
+import com.intellij.patterns.PlatformPatterns
+import com.intellij.psi.xml.XmlFile
+import uk.co.reecedunn.intellij.plugin.core.completion.CompletionContributorEx
+
+class RewriterCompletionContributor : CompletionContributorEx() {
+    @Suppress("PropertyName")
+    val Xml = PlatformPatterns.psiElement().inFile(PlatformPatterns.psiFile(XmlFile::class.java))
+
     init {
-        presentation.isItemTextBold = true
-        presentation.tailText = insertText?.tailText
+        builder(Xml).withFilter(MethodValueTypeFilter).addCompletions(MethodValueTypeProvider)
     }
 }